+++ packages/at_client/lib/src/manager/at_client_manager.dart
@@ -69,8 +69,13 @@ class AtClientManager {
 
     _logger.info("setCurrentAtSign called with atSign $atSign");
     AtUtils.fixAtSign(atSign);
+    Proxies.proxyFallbackEnabled = preference.proxyFallbackEnabled;
     secondaryAddressFinder ??= CacheableSecondaryAddressFinder(
-        preference.rootDomain, preference.rootPort);
+      preference.rootDomain,
+      preference.rootPort,
+      proxies: await Proxies.forDirectory(preference.rootDomain),
+      createSocketTimeout: preference.createSocketTimeout,
+    );
     if (_currentAtClient != null &&
         _currentAtClient?.getCurrentAtSign() == atSign) {
       _logger.info(